Rename layouts folder name to navigation
